fix: support own changelog
This project actually wasn't able to generate its own CHANGELOG.md
because it was getting confused by the title of GitHub Issue #1. This
adds an exception for this case.

In addition to the fix itself, this includes some additional error
logging to make problems easier to debug and also commits the
CHANGELOG.md.

Test: npm run changelog